+++ clearnode/api/channel_v1/interface.go
@@ -69,11 +69,11 @@ const EcdsaSigValidatorType SigValidatorType = "ecdsa"
 
 type MemoryStore interface {
 	// IsAssetSupported checks if a given asset (token) is supported on the specified blockchain.
-	IsAssetSupported(asset, tokenAddress string, blockchainID uint32) (bool, error)
+	IsAssetSupported(asset, tokenAddress string, blockchainID uint64) (bool, error)
 
 	// GetAssetDecimals checks if an asset exists and returns its decimals in YN
 	GetAssetDecimals(asset string) (uint8, error)
 
 	// GetTokenDecimals returns the decimals for a token on a specific blockchain
-	GetTokenDecimals(blockchainID uint32, tokenAddress string) (uint8, error)
+	GetTokenDecimals(blockchainID uint64, tokenAddress string) (uint8, error)
 }
